+++ src/components/alert/test/basic/app/app.component.ts
@@ -1,0 +1,9 @@
+import { Component } from '@angular/core';
+import { RootPage } from '../pages/root-page/root-page';
+
+@Component({
+  template: '<ion-nav [root]="root"></ion-nav>'
+})
+export class E2EApp {
+  root = RootPage;
+}
