+++ src/components/app/test/cordova/app/app.component.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+
+import { PageOne } from '../pages/page-one/page-one';
+
+@Component({
+  templateUrl: `./app.component.html`
+})
+export class E2EApp {
+  root = PageOne;
+}
