+++ src/components/app/test/cordova/app/main.ts
