+++ src/components/app/test/cordova/pages/modal/modal.module.ts
@@ -1,0 +1,17 @@
+import { NgModule } from '@angular/core';
+import { DeepLinkModule } from '../../../../../..';
+
+import { MyModal } from './modal';
+
+@NgModule({
+  declarations: [
+    MyModal
+  ],
+  imports: [
+    DeepLinkModule.forChild(MyModal)
+  ],
+  entryComponents: [
+    MyModal
+  ]
+})
+export class MyModalModule {}
