+++ src/components/app/test/cordova/pages/page-one/provider-one.ts
@@ -1,0 +1,10 @@
+import { Injectable } from '@angular/core';
+
+@Injectable()
+export class SomeData {
+  constructor() {}
+
+  getData() {
+    return 'SomeData';
+  }
+}
