+++ src/components/app/test/cordova/pages/page-one/provider-two.ts
@@ -1,0 +1,10 @@
+import { Injectable } from '@angular/core';
+
+@Injectable()
+export class OtherData {
+  constructor() {}
+
+  getData() {
+    return 'OtherData';
+  }
+}
