+++ src/components/app/test/cordova/pages/page-three/page-three.module.ts
@@ -1,0 +1,17 @@
+import { NgModule } from '@angular/core';
+import { DeepLinkModule } from '../../../../../..';
+
+import { PageThree } from './page-three';
+
+@NgModule({
+  declarations: [
+    PageThree,
+  ],
+  imports: [
+    DeepLinkModule.forChild(PageThree)
+  ],
+  entryComponents: [
+    PageThree,
+  ]
+})
+export class PageThreeModule {}
