+++ src/components/app/test/cordova/pages/page-three/page-three.ts
@@ -1,0 +1,13 @@
+import { Component } from '@angular/core';
+import { NavController } from '../../../../../..';
+
+@Component({
+  templateUrl: 'page-three.html'
+})
+export class PageThree {
+  constructor(public navCtrl: NavController) {}
+
+  goBack() {
+    this.navCtrl.pop();
+  }
+}
