+++ src/components/app/test/cordova/pages/page-two/page-two.module.ts
@@ -1,0 +1,19 @@
+import { NgModule } from '@angular/core';
+import { DeepLinkModule } from '../../../../../..';
+
+import { PageTwo } from './page-two';
+import { MyModalModule } from '../modal/modal.module';
+
+@NgModule({
+  declarations: [
+    PageTwo,
+  ],
+  imports: [
+    DeepLinkModule.forChild(PageTwo),
+    MyModalModule
+  ],
+  entryComponents: [
+    PageTwo,
+  ]
+})
+export class PageTwoModule {}
