+++ src/components/app/test/cordova/pages/tabs-page-one/tabs-page-one.module.ts
@@ -1,0 +1,17 @@
+import { NgModule } from '@angular/core';
+import { DeepLinkModule } from '../../../../../..';
+
+import { TabsPageOne } from './tabs-page-one';
+
+@NgModule({
+  declarations: [
+    TabsPageOne,
+  ],
+  imports: [
+    DeepLinkModule.forChild(TabsPageOne)
+  ],
+  entryComponents: [
+    TabsPageOne,
+  ]
+})
+export class TabsPageOneModule {}
