+++ src/components/app/test/cordova/pages/tabs-page-one/tabs-page-one.ts
@@ -1,0 +1,24 @@
+import { Component } from '@angular/core';
+
+@Component({
+  template: `
+  <ion-header>
+    <ion-navbar>
+      <ion-title>This is a tab page</ion-title>
+      <button ion-button menuToggle>
+        <ion-icon name="menu"></ion-icon>
+      </button>
+      <ion-buttons end>
+        <button ion-button>
+          <ion-icon name="funnel"></ion-icon>
+        </button>
+      </ion-buttons>
+    </ion-navbar>
+  </ion-header>
+  <ion-content padding>
+    <p>The toolbar should have status bar padding.</p>
+  </ion-content>
+  `
+})
+export class TabsPageOne {
+}
