+++ src/components/app/test/cordova/pages/tabs/tabs-page.html
