+++ src/components/app/test/cordova/pages/tabs/tabs-page.module.ts
@@ -1,0 +1,17 @@
+import { NgModule } from '@angular/core';
+import { DeepLinkModule } from '../../../../../..';
+
+import { TabsPage } from './tabs-page';
+
+@NgModule({
+  declarations: [
+    TabsPage,
+  ],
+  imports: [
+    DeepLinkModule.forChild(TabsPage)
+  ],
+  entryComponents: [
+    TabsPage,
+  ]
+})
+export class TabsPageModule {}
