+++ src/components/app/test/gesture-collision/app/main.ts
