+++ src/components/app/test/gesture-collision/pages/page-two/page-two.html
@@ -84,7 +84,7 @@
       <ion-toggle></ion-toggle>
     </ion-item>
 
-    <button ion-item (click)="goToPage1()">Push same page</button>
+    <button ion-item (click)="goToPageTwo()">Push same page</button>
 
     <ion-item>
       <ion-label>RANGE</ion-label>
